Add function comments
